tests/wpt/web-platform-tests/webxr/xrFrame_lifetime.https.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src="resources/webxr_util.js"></script>
  <canvas></canvas>

  <script>
    let immersiveTestName = "XRFrame methods throw exceptions outside of the " +
      "requestAnimationFrame callback for immersive sessions";
    let nonImmersiveTestName = "XRFrame methods throw exceptions outside of the " +
      "requestAnimationFrame callback for non-immersive sessions";

    let fakeDeviceInitParams = { supportsImmersive:true };

    let immersiveSessionOptions = { immersive: true };
    let nonImmersiveSessionOptions = { outputContext: getOutputContext() };

    let testFunction = (testSession, testController, t) => new Promise((resolve) => {
      let staleFrame = null;
      let frameOfRef = null;

      function onFrame(time, xrFrame) {
        t.step(() => {
          assert_true(xrFrame instanceof XRFrame);
        });

        staleFrame = xrFrame;

        step_timeout(afterFrame, 0);
      }

      function afterFrame() {
        t.step(() => {
          // Attempting to call a method on the frame outside the callback that
          // originally provided it should cause it to throw an exception.
          assert_throws('NotAllowedError', () => staleFrame.getDevicePose(frameOfRef));
        });

        // Test does not complete until the this function has executed.
        resolve();
      }

      testSession.requestFrameOfReference('eye-level').then((xrFrameOfRef) => {
        frameOfRef = xrFrameOfRef;

        testSession.requestAnimationFrame(onFrame);
      });
    });

    xr_session_promise_test(immersiveTestName, testFunction,
      fakeDeviceInitParams, immersiveSessionOptions);
    xr_session_promise_test(nonImmersiveTestName, testFunction,
      fakeDeviceInitParams, nonImmersiveSessionOptions);

  </script>
</body>
